implemented aggrigte pipline to counte average of bootcamp courses fees and crud operation on course route

+++ controllers/coursesController.js
@@ -4,6 +4,7 @@
 
 const Course = require('../models/courseModel');
 const asyncHandler = require('../middleware/async');
+const Bootcamp = require('../models/bootcampModel');
 const ErrorResponse = require('../utils/errorResponse');
 const getMany = asyncHandler(async (req, res, next) => {
 	let query;
@@ -19,7 +20,16 @@ const getMany = asyncHandler(async (req, res, next) => {
 		data   : courses
 	});
 });
-const posOne = asyncHandler(async (req, res, next) => {
+//Add course
+//route POST /api/v1/bootcamp/:bootcampId/courses
+//access Privat
+const postOne = asyncHandler(async (req, res, next) => {
+	const id = req.params.bootcampId;
+	req.body.bootcamp = id;
+	const bootcamp = await Bootcamp.findById(id).exec();
+	if (!bootcamp) {
+		return next(new ErrorResponse(`There is no Bootcamp with an Id ${id}`));
+	}
 	const item = await Course.create(req.body);
 	res.status(200).json({
 		status : 'Success',
@@ -27,14 +37,22 @@ const posOne = asyncHandler(async (req, res, next) => {
 	});
 });
 const getOne = asyncHandler(async (req, res, next) => {
-	const item = await Course.findById(req.params.id);
+	const item = await Course.findById(req.params.id).exec();
+
+	if (!item) {
+		return next(new ErrorResponse(`No course with id of ${req.params.id}`, 404));
+	}
 	res.status(200).json({
 		status : 'Success',
 		data   : item
 	});
 });
 const updateOne = asyncHandler(async (req, res, next) => {
-	const item = await Course.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: false });
+	let item = await Course.findById(req.params.id);
+	if (!item) {
+		return next(new ErrorResponse(`No course with id of ${req.params.id}`, 404));
+	}
+	item = await Course.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: false });
 	res.status(200).json({
 		status : 'Success',
 		data   : item
@@ -42,6 +60,10 @@ const updateOne = asyncHandler(async (req, res, next) => {
 });
 
 const deleteOne = asyncHandler(async (req, res, next) => {
+	let item = await Course.findById(req.params.id);
+	if (!item) {
+		return next(new ErrorResponse(`No course with id of ${req.params.id}`, 404));
+	}
 	await Course.findByIdAndRemove(req.params.id);
 	res.status(204).json({
 		status : 'Success',
@@ -50,7 +72,7 @@ const deleteOne = asyncHandler(async (req, res, next) => {
 });
 module.exports = {
 	getMany,
-	posOne,
+	postOne,
 	getOne,
 	deleteOne,
 	updateOne
